move download settings to own PreferencePane
